vim-patch:8.1.1927: code for dealing with script files is spread out
Problem:    Code for dealing with script files is spread out.
Solution:   Move the code to scriptfile.c. (Yegappan Lakshmanan, closes vim/vim#4861)
https://github.com/vim/vim/commit/307c5a5bb77c3728dfab06c30e9f786309c63f74
